+++ flagshiptut/src/main/java/org/moonwave/jpa/model/EvaluationObjective.java
@@ -22,7 +22,7 @@ public class EvaluationObjective implements Serializable {
 
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private int id;
+    private Integer id;
 
     @Column(name="create_time")
     private Timestamp createTime;
@@ -79,11 +79,11 @@ public class EvaluationObjective implements Serializable {
     public EvaluationObjective() {
     }
 
-    public int getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public void setId(int id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -233,7 +233,7 @@ public class EvaluationObjective implements Serializable {
 
     @Override
     public int hashCode() {
-        return ((Integer)id).hashCode();
+        return id.hashCode();
     }
 
     @Override
