+++ flagshiptut/src/main/java/org/moonwave/jpa/model/Semester.java
@@ -22,7 +22,7 @@ public class Semester implements Serializable {
 
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private short id;
+    private Short id;
 
     private String alias;
 
@@ -45,11 +45,11 @@ public class Semester implements Serializable {
     public Semester() {
     }
 
-    public short getId() {
+    public Short getId() {
         return this.id;
     }
 
-    public void setId(short id) {
+    public void setId(Short id) {
         this.id = id;
     }
 
@@ -103,7 +103,7 @@ public class Semester implements Serializable {
 
     @Override
     public int hashCode() {
-        return ((Short)id).hashCode();
+        return id.hashCode();
     }
 
     @Override
